Added Request and Handler for Product.

+++ src/Domain/Entities/Entity.cs
@@ -2,7 +2,7 @@
 {
     public abstract class Entity
     {
-        public virtual long Id { get; protected set; }
+        public virtual int Id { get; protected set; }
 
         public override bool Equals(object obj)
         {
